--- report-service/src/main/java/gov/cdc/dataingestion/deadletter/repository/model/DeadLetterELRModel.java
@@ -1,124 +1,0 @@
-package gov.cdc.dataingestion.deadletter.repository.model;
-
-import jakarta.persistence.*;
-import org.hibernate.annotations.GenericGenerator;
-
-import java.sql.Timestamp;
-
-@Entity
-@Table(name = "elr_dlt")
-public class DeadLetterELRModel {
-    @Id
-    @GenericGenerator(name = "generator", strategy = "guid", parameters = {})
-    @GeneratedValue(generator = "generator")
-    @Column(name = "id" , columnDefinition="uniqueidentifier")
-    private String id;
-
-    @Column(name = "error_message_id")
-    private String rawId;
-
-    @Column(name = "error_message_source")
-    private String rawMessage;
-
-    @Column(name = "error_stack_trace")
-    private String errorStackTrace;
-
-    @Column(name = "dlt_occurrence")
-    private Integer dltOccurrence;
-
-    @Column(name="dlt_status")
-    private String dltStatus;
-
-    @Transient
-    @Column(name = "created_on")
-    private Timestamp createdOn;
-
-    @Column(name = "updated_on")
-    private Timestamp updatedOn;
-
-    @Column(name = "created_by")
-    private String createdBy;
-
-    @Column(name = "updated_by")
-    private String updatedBy;
-
-    public String getDltStatus() {
-        return dltStatus;
-    }
-
-    public void setDltStatus(String dltStatus) {
-        this.dltStatus = dltStatus;
-    }
-
-    public Integer getDltOccurrence() {
-        return dltOccurrence;
-    }
-
-    public void setDltOccurrence(Integer dltOccurrence) {
-        this.dltOccurrence = dltOccurrence;
-    }
-
-    public String getErrorStackTrace() {
-        return errorStackTrace;
-    }
-
-    public void setErrorStackTrace(String errorStackTrace) {
-        this.errorStackTrace = errorStackTrace;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getRawId() {
-        return rawId;
-    }
-
-    public void setRawId(String rawId) {
-        this.rawId = rawId;
-    }
-
-    public String getRawMessage() {
-        return rawMessage;
-    }
-
-    public void setRawMessage(String rawMessage) {
-        this.rawMessage = rawMessage;
-    }
-
-    public Timestamp getCreatedOn() {
-        return createdOn;
-    }
-
-    public void setCreatedOn(Timestamp createdOn) {
-        this.createdOn = createdOn;
-    }
-
-    public Timestamp getUpdatedOn() {
-        return updatedOn;
-    }
-
-    public void setUpdatedOn(Timestamp updatedOn) {
-        this.updatedOn = updatedOn;
-    }
-
-    public String getCreatedBy() {
-        return createdBy;
-    }
-
-    public void setCreatedBy(String createdBy) {
-        this.createdBy = createdBy;
-    }
-
-    public String getUpdatedBy() {
-        return updatedBy;
-    }
-
-    public void setUpdatedBy(String updatedBy) {
-        this.updatedBy = updatedBy;
-    }
-}
